[Fixed] Changes Query Namespace
