+++ PersonalFinanceManager.Query/PersonalFinanceManager.Query.csproj
@@ -41,9 +41,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AddNewData\SqlQuery.cs" />
-    <Compile Include="Budget\SqlQuery.cs" />
+    <Compile Include="SqlQuery.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
